+++ include/roq/cache/market_by_order.hpp
@@ -35,8 +35,8 @@ struct ROQ_PUBLIC MarketByOrder {
   virtual double quantity_increment() const = 0;
 
   // precision (required to show all significant decimal places)
-  virtual Precision price_decimals() const = 0;
-  virtual Precision quantity_decimals() const = 0;
+  virtual Precision price_precision() const = 0;
+  virtual Precision quantity_precision() const = 0;
 
   // last update
   virtual uint16_t stream_id() const = 0;